+++ timeline/tests/test_permissions.py
@@ -39,8 +39,28 @@ class TestTimelinePermissions(TestProjectPermissionBase):
         self.project.set_public()
         self.assert_response(url, self.anonymous, 200)
 
-    def test_object_list(self):
-        """Test object event list"""
+    def test_site_list(self):
+        """Test site event list"""
+        url = reverse('timeline:list_site')
+        good_users = [
+            self.superuser,
+            self.owner_as.user,
+            self.delegate_as.user,
+            self.contributor_as.user,
+            self.guest_as.user,
+            self.user_no_roles,
+        ]
+        self.assert_response(url, good_users, 200)
+        self.assert_response(url, self.anonymous, 302)
+
+    @override_settings(PROJECTROLES_ALLOW_ANONYMOUS=True)
+    def test_stite_list_anon(self):
+        """Test site event list with anonynomus access"""
+        url = reverse('timeline:list_site')
+        self.assert_response(url, self.anonymous, 302)
+
+    def test_project_event_object_list(self):
+        """Test project event object list"""
         url = reverse(
             'timeline:list_object',
             kwargs={
@@ -65,7 +85,7 @@ class TestTimelinePermissions(TestProjectPermissionBase):
         self.assert_response(url, self.anonymous, 302)
 
     @override_settings(PROJECTROLES_ALLOW_ANONYMOUS=True)
-    def test_object_list_anon(self):
+    def test_project_event_object_list_anon(self):
         """Test object event list with anonynomus access"""
         url = reverse(
             'timeline:list_object',
@@ -77,3 +97,40 @@ class TestTimelinePermissions(TestProjectPermissionBase):
         )
         self.project.set_public()
         self.assert_response(url, self.anonymous, 200)
+
+    def test_site_event_object_list(self):
+        """Test site event object list"""
+        url = reverse(
+            'timeline:list_object_site',
+            kwargs={
+                'object_model': 'User',
+                'object_uuid': self.user_owner.sodar_uuid,
+            },
+        )
+        good_users = [
+            self.superuser,
+            self.owner_as.user,
+            self.delegate_as.user,
+            self.contributor_as.user,
+            self.guest_as.user,
+            self.user_no_roles,
+        ]
+        self.assert_response(url, good_users, 200)
+        self.assert_response(url, self.anonymous, 302)
+        # Test public project
+        self.project.set_public()
+        self.assert_response(url, self.user_no_roles, 200)
+        self.assert_response(url, self.anonymous, 302)
+
+    @override_settings(PROJECTROLES_ALLOW_ANONYMOUS=True)
+    def test_site_event_object_list_anon(self):
+        """Test site event list with anonymous access"""
+        url = reverse(
+            'timeline:list_object_site',
+            kwargs={
+                'object_model': 'User',
+                'object_uuid': self.user_owner.sodar_uuid,
+            },
+        )
+        self.project.set_public()
+        self.assert_response(url, self.anonymous, 302)